h3i::actions
Module h3

Actions specific to HTTP/3 and QUIC

Actions are small operations such as sending HTTP/3 frames or managing QUIC streams. Each independent use case for h3i requires its own collection of Actions, that h3i iterates over in sequence and executes.

Structs

StreamEvent

A response event, received over a stream, which will terminate the wait period.

Enums

Action

An action which the HTTP/3 client should take.

StreamEventType

Response that can terminate a wait period.

WaitType

Configure the wait behavior for a connection.

Functions

send_headers_frame

Convenience to convert between header-related data and a Action::SendHeadersFrame.