h3i
Module client

The main h3i client runner.

The client is responsible for connecting to an indicated server, executing as series of Actions, and capturing the results in a ConnectionSummary.

Modules

connection_summary

Summarizes events that occurred during a connection.

sync_client

Responsible for creating a quiche::Connection and managing I/O.

Enums

ClientError

Represents different errors that can occur when sync_client runs.