h3i::client

Module connection_summary

Summarizes events that occurred during a connection.

Structs

CloseTriggerFrames

A container for frames that h3i expects to see over a given connection. If h3i receives all the frames it expects, it will send a CONNECTION_CLOSE frame to the server. This bypasses the idle timeout and vastly quickens test suites which depend heavily on h3i.

ConnectionCloseDetails

Denotes why the connection was closed.

ConnectionSummary

A summary of all frames received on a connection. There are some extra fields included to provide additional context into the connection’s behavior.

SerializableConnectionError

A wrapper to help serialize a quiche::ConnectionError

SerializablePathStats

A wrapper to help serialize quiche::PathStats

SerializableStats

A wrapper to help serialize quiche::Stats

StreamMap

A read-only aggregation of frames received over a connection, mapped to the stream ID over which they were received.

Constants

MAX_SERIALIZED_BUFFER_LEN

Maximum length of any serialized element’s unstructured data such as reason phrase.