h3i::client::connection_summary

Struct StreamMap

pub struct StreamMap { /* private fields */ }

A read-only aggregation of frames received over a connection, mapped to the stream ID over which they were received.

StreamMap also contains the CloseTriggerFrames for the connection so that its state can be updated as new frames are received.

Implementations

impl StreamMap

pub fn all_frames(&self) -> Vec<H3iFrame>

Flatten all received frames into a single vector. The ordering is non-deterministic.

Example

use h3i::client::connection_summary::StreamMap;
use h3i::frame::EnrichedHeaders;
use h3i::frame::H3iFrame;
use quiche::h3::Header;
use std::iter::FromIterator;

let h = Header::new(b"hello", b"world");
let headers = H3iFrame::Headers(EnrichedHeaders::from(vec![h]));

let stream_map: StreamMap = [(0, vec![headers.clone()])].into();
assert_eq!(stream_map.all_frames(), vec![headers]);

pub fn stream(&self, stream_id: u64) -> Vec<H3iFrame>

Get all frames on a given stream_id.

Example

use h3i::client::connection_summary::StreamMap;
use h3i::frame::EnrichedHeaders;
use h3i::frame::H3iFrame;
use quiche::h3::Header;
use std::iter::FromIterator;

let mut stream_map = StreamMap::default();

let h = Header::new(b"hello", b"world");
let headers = H3iFrame::Headers(EnrichedHeaders::from(vec![h]));

let stream_map: StreamMap = [(0, vec![headers.clone()])].into();
assert_eq!(stream_map.stream(0), vec![headers]);

pub fn received_frame(&self, frame: &H3iFrame) -> bool

Check if a provided H3iFrame was received, regardless of what stream it was received on.

Example

use h3i::client::connection_summary::StreamMap;
use h3i::frame::EnrichedHeaders;
use h3i::frame::H3iFrame;
use quiche::h3::Header;
use std::iter::FromIterator;

let h = Header::new(b"hello", b"world");
let headers = H3iFrame::Headers(EnrichedHeaders::from(vec![h]));

let stream_map: StreamMap = [(0, vec![headers.clone()])].into();
assert!(stream_map.received_frame(&headers));

pub fn received_frame_on_stream(&self, stream: u64, frame: &H3iFrame) -> bool

Check if a provided H3iFrame was received over a specified stream.

Example

use h3i::client::connection_summary::StreamMap;
use h3i::frame::EnrichedHeaders;
use h3i::frame::H3iFrame;
use quiche::h3::Header;
use std::iter::FromIterator;

let h = Header::new(b"hello", b"world");
let headers = H3iFrame::Headers(EnrichedHeaders::from(vec![h]));

let stream_map: StreamMap = [(0, vec![headers.clone()])].into();
assert!(stream_map.received_frame_on_stream(0, &headers));

pub fn is_empty(&self) -> bool

Check if the stream map is empty, e.g., no frames were received.

Example

use h3i::client::connection_summary::StreamMap;
use h3i::frame::EnrichedHeaders;
use h3i::frame::H3iFrame;
use quiche::h3::Header;
use std::iter::FromIterator;

let mut stream_map = StreamMap::default();
assert!(stream_map.is_empty());

let h = Header::new(b"hello", b"world");
let headers = H3iFrame::Headers(EnrichedHeaders::from(vec![h]));

let stream_map: StreamMap = [(0, vec![headers.clone()])].into();
assert!(!stream_map.is_empty());

pub fn headers_on_stream(&self, stream_id: u64) -> Vec<EnrichedHeaders>

See all HEADERS received on a given stream.

Example

use h3i::client::connection_summary::StreamMap;
use h3i::frame::EnrichedHeaders;
use h3i::frame::H3iFrame;
use quiche::h3::Header;
use std::iter::FromIterator;

let h = Header::new(b"hello", b"world");
let enriched = EnrichedHeaders::from(vec![h]);
let headers = H3iFrame::Headers(enriched.clone());
let data = H3iFrame::QuicheH3(quiche::h3::frame::Frame::Data {
    payload: b"hello world".to_vec(),
});

let stream_map: StreamMap = [(0, vec![headers.clone(), data.clone()])].into();
assert_eq!(stream_map.headers_on_stream(0), vec![enriched]);

pub fn all_close_trigger_frames_seen(&self) -> bool

If all CloseTriggerFrames were seen. If no triggers were expected, this will return false.

pub fn missing_close_trigger_frames(&self) -> Option<Vec<CloseTriggerFrame>>

The set of all CloseTriggerFrames that were not seen on the connection. Returns None if