h3i::config

Struct Config

pub struct Config {

    pub host_port: String,
    pub omit_sni: bool,
    pub connect_to: Option<String>,
    pub source_port: u32,
    pub verify_peer: bool,
    pub idle_timeout: u64,
    pub max_data: u64,
    pub max_stream_data_bidi_local: u64,
    pub max_stream_data_bidi_remote: u64,
    pub max_stream_data_uni: u64,
    pub max_streams_bidi: u64,
    pub max_streams_uni: u64,
    pub max_window: u64,
    pub max_stream_window: u64,
    pub session: Option<Vec<u8>>,
    pub enable_dgram: bool,
    pub dgram_recv_queue_len: usize,
    pub dgram_send_queue_len: usize,

}

Server details and QUIC connection properties.

Fields
host_port: String

A string representing the host and port to connect to using the format <host>:<port>.

omit_sni: bool

If the SNI should be omitted during the TLS handshake.

connect_to: Option<String>

Set a specific IP address to connect to, rather than use DNS resolution.

source_port: u32

The source port to use when connecting to a server.

verify_peer: bool

Whether to verify the server certificate.

idle_timeout: u64

The QUIC idle timeout value in milliseconds.

max_data: u64

Flow control limit for the connection in bytes

max_stream_data_bidi_local: u64

Flow control limit for locally-initiated bidirectional streams in bytes.

max_stream_data_bidi_remote: u64

Flow control limit for remotely-initiated bidirectional streams in bytes.

max_stream_data_uni: u64

Flow control limit for unidirectional streams in bytes.

max_streams_bidi: u64

Maximum count for concurrent remotely-initiated bidirectional streams.

max_streams_uni: u64

โMaximum count for concurrent remotely-initiated unidirectional streamsโ.

max_window: u64

Receiver window limit for the connection in bytes.

max_stream_window: u64

Receiver window limit for a stream in bytes.

session: Option<Vec<u8>>

Set the session to attempt resumption.

enable_dgram: bool

Whether to enable datagram sending.

dgram_recv_queue_len: usize

Datagram receive queue length.

dgram_send_queue_len: usize

Datagram send queue length.

Implementations

impl Config

pub fn new() -> Self

Construct a new config object with default values.

pub fn with_host_port(self, host_port: String) -> Self

pub fn omit_sni(self) -> Self

pub fn with_connect_to(self, connect_to: String) -> Self

pub fn with_source_port(self, port: u32) -> Self

pub fn verify_peer(self, verify_peer: bool) -> Self

pub fn with_idle_timeout(self, idle_timeout: u64) -> Self

pub fn with_max_data(self, max_data: u64) -> Self

pub fn with_max_stream_data_bidi_local( self, max_stream_data_bidi_local: u64, ) -> Self

pub fn with_max_stream_data_bidi_remote( self, max_stream_data_bidi_remote: u64, ) -> Self

pub fn with_max_stream_data_uni(self, max_stream_data_uni: u64) -> Self

pub fn with_max_streams_bidi(self, max_streams_bidi: u64) -> Self

pub fn with_max_streams_uni(self, max_streams_uni: u64) -> Self

pub fn with_max_window(self, max_window: u64) -> Self

pub fn with_max_stream_window(self, max_stream_window: u64) -> Self

pub fn enable_dgram(self, enable_dgram: bool) -> Self

pub fn with_dgram_recv_queue_len(self, dgram_recv_queue_len: usize) -> Self

pub fn with_dgram_send_queue_len(self, dgram_send_queue_len: usize) -> Self

pub fn build(self) -> Result<Self, Error>