h3i
Module frame

Helpers for dealing with quiche stream events and HTTP/3 frames.

Structs

EnrichedHeaders

An HTTP/3 HEADERS frame with decoded headers and a HeaderMap.

ResetStream

A RESET_STREAM frame.

SerializableHeader

A wrapper to help serialize an quiche HTTP header.

SerializableQFrame

A wrapper to help serialize a quiche HTTP/3 frame.

Enums

H3iFrame

An internal representation of a QUIC or HTTP/3 frame. This type exists so that we can extend types defined in Quiche.

Type Aliases

BoxError