Crate h3i

h3i - low-level HTTP/3 debug and testing

HTTP/3 (RFC 9114) is the wire format for HTTP semantics (RFC 9110). The RFCs contain a range of requirements about how Request or Response messages are generated, serialized, sent, received, parsed, and consumed. QUIC (RFC 9000) streams are used for these messages along with other control and QPACK (RFC 9204) header compression instructions.

h3i provides a highly configurable HTTP/3 client that can bend RFC rules in order to test the behavior of servers. QUIC streams can be opened, fin’d, stopped or reset at any point in time. HTTP/3 frames can be sent on any stream, in any order, containing user-controlled content (both legal and illegal).

Re-exports

pub use quiche;

Modules

actions

Actions are small operations such as sending HTTP/3 frames or managing QUIC streams. Each independent use case for h3i requires its own collection of Actions, that h3i iterates over in sequence and executes.

client

The main h3i client runner.

config

Configuration for the h3i client.

frame

Helpers for dealing with quiche stream events and HTTP/3 frames.

frame_parser

Stateful parsing of QUIC streams into HTTP/3 frames.

prompts

A collection of interactive CLI prompts based on [inquire].

recordreplay

Support for recording h3i Actions and replaying them.

Constants

HTTP3_CONTROL_STREAM_TYPE_ID

The ID for an HTTP/3 control stream type.

HTTP3_PUSH_STREAM_TYPE_ID

The ID for an HTTP/3 push stream type.

QPACK_DECODER_STREAM_TYPE_ID

The ID for a QPACK decoder stream type.

QPACK_ENCODER_STREAM_TYPE_ID

The ID for a QPACK encoder stream type.