h3i

Module client

The main h3i client runner.

The client is responsible for connecting to an indicated server, executing as series of Actions, and capturing the results in a ConnectionSummary.

Modules

async_client

Responsible for creating a tokio_quiche::quic::QuicheConnection and yielding I/O to tokio-quiche.

connection_summary

Summarizes events that occurred during a connection.

sync_client

Responsible for creating a quiche::Connection and managing I/O.

Enums

ClientError

Represents different errors that can occur when the h3i client runs.